compilation warnings fixed
git-svn-id: https://vcs.maemo.org/svn/min/trunk@127 86ee150f-fb1f-4fea-b4f8-59f22ff7da1d

+++ min/src/test_libraries/tcCases.c
@@ -126,7 +126,7 @@ LOCAL int FailingOne (MinItemParser * item)
 LOCAL int CheckThreshold (MinItemParser * item)
 {
         int threshold;
-        int failcount;
+        long failcount;
 
         if (mip_get_next_int (item, &threshold) != ENOERR) {
                 tm_printf (0, "", "Threshold parameter missing");
